Fix libarary -> library typo
We are over half a century in to this programming thing; text editing and spell checkers will continue to be an unsolved problem next century too.

---

Thanks for all the awesome work, noticed this while reading through. The documentation you all have is fantastic. Looking forward to exploring integrating this.