Fix lock semantics for the exclusive connection lock when running against a pooler (#2791)
We have recently observed some source in Electric Cloud struggle due to
lock and replication slot contention issues which in theory shouldn't
have happened at all. The cause of the issues was an imperfect
validation logic applied to the creation of new sources. As a
consequence, some sources get created with a pooled connection URL.

More concretely, our validation can successfully weed out pooled
connection strings from Supabase, but Neon's pool allows establishing
replication connections by proxying them directly to the DB. As a
result, Electric ends up acquiring an advisory lock on a pooled
connection where the pooler runs in transaction mode, leading to a bunch
of undesirable effects that prevent Electric's normal operation.

---

This PR addresses the cause of the issues by making the `LockConnection`
process open a replication connection to the DB or fail if that cannot
be achieved. In this way, either the advisory lock will be taken on a
long-living connection as expected, or the connection won't be able to
open and the source will immediately transition into an error state,
instead of assigning an advisory lock to a random pooled connection as
it currently does.