fix(elixir-client): Add handling of Ecto.ULID, :map and :array types (#2815)
ULID fields expect to `cast/1` a base-32 encoded value but since the
backing column type is uuid, the values in the replication stream are
the string representation of a uuid, so trying to cast these raises.
This pr adds special handling for custom types that use UUID-type
columns and attempt to simulate a load of a raw uuid value from the db.

Fixes https://github.com/electric-sql/phoenix_sync/issues/24