Fix flaky ShapeCache test
Before this change, it was easy to see the test fail after just a
handful of runs. The reason was that the simulated snapshotter process
would be resumed before the test process had started waiting on the
snapshot.

This change makes sure there's a process in the consumer's waiting list
before resuming the simulated snapshotter, though it does make the test
aware of implementation details.