fix: Use separate stack event for pool connection errors (#2864)
Fixes https://github.com/electric-sql/stratovolt/issues/697

We need to classify these pool connection errors differently than the
errors that cause either a reconnection or a failure to electric.

We can do that either with a separate stack event as suggested by @alco
and done here with `:pool_connection_error`, or adding a property to
indicate that we can ignore this. I am of the opinion that this should
be a clearly different event since it is not fatal or causes a full
reconnection, but am happy to hear other thoughts - my understanding is
the pool connections will be reinstated?