fix: Ensure PG expression execution in Electric matches PG (#2862)
Introduces some `StreamData` property tests for our PG expression parser
and executor.

Not the prettiest but I did already find several issues, some of which I
would need confirmation that they were not intentional:

1. `AND` should handle `false and null` case as `false`
2. Array concatenation with `null` should return the same array
3. `array_prepend` and `array_append` work with `null`s, and they
actually add the null to the array
4. Array prepends and appends using `||` work with `null` but they keep
the array the same, and they do not work with strings
5. Array contains and is contained have specific behaviours with `null`
- if the array that should be contained has a `null` then it will never
pass the check - whether the array containing the other has a `null` or
not.
6. The `^` operator should handle `0^0`, which `Float.pow` did not - not
sure if this affects performance significantly @icehaunter
7. Added bitwise not operator - not necessary, can remove, was just
checking most operators


I've also found that various comparison operators don't work on booleans
and arrays, but decided to leave that as separate work and instead try
to get this merged as it is an improvement over what we had before.

The tests generate arbitrary expressions and then use postgres as the
oracle to ensure that the values we produce are the same.

If both our runner and postgres error, we ignore the result.

If our runner resolves to nil but postgres errors I also ignore that
result, as we have this concept of `strict?` execution that would need
examination in lots of parts to ensure we have the same behaviour as
postgres exactly. Can be done iteratively.