fix: Parse `NULL`s correctly in TS client (#2882)
Fixes https://github.com/electric-sql/electric/issues/2877

We used to parse tokens of the form `NULL` to `null` in TS as we needed
to pars the nulls within arrays (that are received as string tokens).
However this ended up causing any text value `NULL` to be parsed as a
`null` too.

In order to fix this we directly convert the `NULL` string tokens to
`null` inside `pgArrayParser`, which can handle quoted `"NULL"` within
arrays correctly so that it can pass it on to subsequent parsing stages
as a string.