fix: Ensure telemetry handlers are detached appropriately (#2884)
Closes https://github.com/electric-sql/electric/issues/2768

In the `OtelMetricExporter`, the handlers were not getting explicitly
detached upon shutting down the relevant processes, and thus we were
seeing errors when the handlers were getting executed by the poller
after the process had been removed. That has been fixed in the relevant
library and I have updated the version here, see
https://github.com/electric-sql/elixir-otel-metric-exporter/pull/12.

In the `CallHomeReporter`, although we were calling `:telemetry.detach`
explicitly, we were not using the actual handler IDs and thus we were
not actually removing the handlers.