When a shape consumer stops, it doesn't need to notify snapshot waiters
...presumbly?

Context:

If the publication manager finds an error during shape creation, it
sends it back to the snapshotter process and also invalidates the shapes
for changed relations. The problem with the latter is that it
immediately stops the shape consumer and sends a must_refetch() message
to the API request process, before the snapshotter is able to process
the error from the publication manager. So the API request gets a
generic 409 response instead of the specific error from the publication
manager.

As a matter of fact, it's not necessary for the shape consumer to notify
snapshot waiters since the snapshotter process already does that when it
encounters an error.