chore: Refactor `ShapeStatus` table to begin separately before `ShapeCache` and consumers (#3012)
This work precedes the work I'm doing on
https://github.com/electric-sql/electric/issues/2984

The `ShapeStatus` ETS table was being initialised by `ShapeCache`, but
consumers also use it, and the consumer supervisor lives before the
shape cache, causing issues in tests during shutdowns etc.

This changes _no_ behaviour in Electric (or at least it shouldn't!).
I've refactored the tests and everything relevant to use shape cache in
its new form.

I still don't love how it looks and I think it can be simplified further
but this addresses the initial concern of moving it up front, and
simplifies it as a mod arg passed at the stack supervisor level and then
passed as is to appropriate modules, with a dedicated `ShapeStatusAgent`
initializing and owning the table.