fix: Consistent cache busting on 409s (#3021)
Fixes https://github.com/electric-sql/electric/issues/3020

Before adapting the cursor for the purpose of 409s, I decided to use the
handle itself which we were already using.

Here is my reasoninig:

1. Our main issue is that we are caching 409s (with or without a next
handle) for 60 seconds - we are caching them as they are part of our
regular request flow and it would create unnecessary load to break
request coalescing and go to origin for them
1a. One immediate fix is to change the caching for 409s _without_ a
handle to 1 second, the minimum to enable request coalescing. This means
that on the next request cycle they would most likely run into a 409
_with_ a shape handle anyway.
1b. We could disable caching on 409s without a handle altogether - I
_think_ that would would happen is that the 409 is return to one of the
"held" requests, and the next one is picked as the "champion", and that
one goes to the origin and it would _now_ receive a 409 _with_ a shape
handle - so perhaps this is the better option.
2. I'm explicitly removing the shape handle as part of the `offset=-1`
requests, so it really now only acts as a cache buster and it is not
respected.
2a. This allows us to "abuse" the `handle` parameter for `offset=-1`
requests as a cache buster (which we were already sort of doing) - and
rather than rely on a cursor we can make clients just request the same
handle with a `-next` suffix, which is already consistent as per
@samwillis ' suggestion and does not require coordination.


The reason I'm avoiding reusing the cursor is that it is used very
explicitly in clients for _live_ requests, and ultimately it wouldn't
give us a big benefit since all we really want is for clients to go on
to the "next" chain of requests from the one they are on - whether
that's via an explicit new handle that is given or via a consistent
cache buster without coordination is irrelevant. I'd rather not add more
complexity to the protocol.


## IMPORTANT

This change basically makes the `offset=-1` calls _ignore_ the `handle`
parameter, so if you provide an invalid handle we will still match to an
existing shape and return the valid one. If this is not how we would
like the API to behave then I will scrap this change and use a different
approach (with the cursor).