fix: Handle missing publication errors (#3044)
Closes https://github.com/electric-sql/stratovolt/issues/761

The errors we observed were because the publication created by Electric
was being manually deleted by the developer. We should be safe to parse
it into an error that restarts, ensuring consistency.

## EDIT
I've converted it to `retry_may_fix?: true` as 1) that was the current
behaviour and 2) since this error basically restarts the replication
client and the process of creating the slot/publication, this might be
more appropriate?

@alco what do you think? should we force a harder restart on this error
or rely on `retry_may_fix?`?


## EDIT 2

We are now capturing the missing publication error in
`PublicationManager` as well and immediately terminating it, which
causes a subsequent termination of `ShapeLogCollector` and thus a
restart of the replication client which fixes the missing publication.