fix: Pool connections attempting to link to dead process (#3052)
Fixes https://github.com/electric-sql/electric/issues/3049

The error that is hidden is a `:noproc` `ErlangError` when trying to do
`Process.link` from the connection being set up to a supervisor that is
no longer alive.

The underlying DBConnection application has various processes living in
its own extra application handling connection lifecycles with some
levels of indirection so we cannot rely on the pool being terminated to
ensure there's no new connections popping up when the supervisor is
dead.

Instead, we check if the supervisor is alive at the time of the
connection being configured, and if it is not we exit.

I've reproduced the error manually by randomly killing off the
supervisor with timeouts while the pool connections are setting up and
was consistently reproducing this - the suggested fix does the trick. I
was also thinking of rescuing the `ErlangError` but the `alive?` check
feels more explicit (even if it technically has a gap between the check
and the link op?). I don't _love_ it but it does take care of the issue
and I can't find a better approach currently. **EDIT:** I reverted back
to a `try/catch` call - better keep it solid, clarity isn't too much of
an issue


While running tests I also noticed some errors that I figured were
because of an incorrect order of processes in our replication
supervisor. The consumer supervisor should be started after the shape
status table, the shape log collector, and the publication manager, as
consumers make use of all three during their lifetime, so during
controlled shutdowns we want this order to be maintained.